+++ src/data_sources/http/filter/request_body/mod.rs
@@ -13,7 +13,8 @@ impl HttpDataSource {
 
         match resolver_type {
             ResolverType::CreateOne => {
-                let json = Json::from(input);
+                let values = input.get("values").unwrap().as_document().unwrap().clone();
+                let json = Json::from(values);
                 Some(json)
             }
             ResolverType::FindOne | ResolverType::FindMany => None,