+++ src/data_sources/http/services/mod.rs
@@ -1,6 +1,6 @@
 use async_graphql::ErrorExtensions;
 use http::Method;
-use log::{debug, info};
+use log::{debug, error, trace};
 use reqwest::Client;
 
 use super::filter::HttpDataSourceFilter;
@@ -18,23 +18,23 @@ impl Services {
         client: Client,
         filter: HttpDataSourceFilter,
     ) -> Result<String, async_graphql::Error> {
-        info!("Executing Request - HTTP Data Source");
+        debug!("Executing Request - HTTP Data Source");
 
         let response = match filter.method {
             Method::GET => {
-                debug!("Using GET Method");
+                trace!("Using GET Method");
                 let res = client.get(filter.url).send().await?;
                 match res.status().is_success() {
                     true => res.text().await?,
                     _ => {
-                        debug!("Response Status: {:?}", res.status());
+                        error!("Response Status: {:?}", res.status());
                         Err(async_graphql::Error::new("HTTP Request Failed")
                             .extend_with(|_, e| e.set("status", res.status().as_u16())))?
                     }
                 }
             }
             Method::POST => {
-                debug!("Using POST Method");
+                trace!("Using POST Method");
                 let res = client
                     .post(filter.url)
                     .json(&filter.request_body)
@@ -43,13 +43,13 @@ impl Services {
                 match res.status().is_success() {
                     true => res.text().await?,
                     _ => {
-                        debug!("Response Status: {:?}", res.status());
+                        error!("Response Status: {:?}", res.text().await?);
                         Err(async_graphql::Error::new("HTTP Request Failed"))?
                     }
                 }
             }
             Method::PUT => {
-                debug!("Using PUT Method");
+                trace!("Using PUT Method");
                 let res = client
                     .put(filter.url)
                     .json(&filter.request_body)
@@ -58,13 +58,13 @@ impl Services {
                 match res.status().is_success() {
                     true => res.text().await?,
                     _ => {
-                        debug!("Response Status: {:?}", res.status());
+                        error!("Response Status: {:?}", res.status());
                         Err(async_graphql::Error::new("HTTP Request Failed"))?
                     }
                 }
             }
             Method::PATCH => {
-                debug!("Using PATCH Method");
+                trace!("Using PATCH Method");
                 let res = client
                     .patch(filter.url)
                     .json(&filter.request_body)
@@ -73,16 +73,17 @@ impl Services {
                 match res.status().is_success() {
                     true => res.text().await?,
                     _ => {
-                        debug!("Response Status: {:?}", res.status());
+                        error!("Response Status: {:?}", res.status());
                         Err(async_graphql::Error::new("HTTP Request Failed"))?
                     }
                 }
             }
             _ => {
-                debug!("Unsupported Method");
+                error!("Unsupported Method");
                 Err(async_graphql::Error::new("Unsupported Method"))?
             }
         };
+        trace!("Response: {:?}", response);
         Ok(response)
     }
 }